users can only delete or edit their own news items

+++ src/scripts/profile/newsEventListener.js
@@ -7,6 +7,8 @@ import editNewsForm from "./editNewsForm"
 const newsEventListener = () => {
     document.querySelector("#news").addEventListener("click", (event) => {
         let userId = document.querySelector("#userId").value
+        let buttonParent = event.target.parentElement
+
         // ADD NEWS ITEM button
         if (event.target.id === "addNews") {
             let newsFormHTML = newsForm()
@@ -30,49 +32,57 @@ const newsEventListener = () => {
 
         // DELETE button
         } else if (event.target.id.startsWith("deleteNews")) {
-            // Find ID of news item to remove
-            let idToDelete = event.target.id.split("--")[1]
-            let stringToDelete = `news/${idToDelete}`
-            // Delete from API
-            API.DELETE(stringToDelete)
-                // Re-load news list
-                .then(response => {
-                    createNews()
-                })
 
-        // EDIT button
-        } else if (event.target.id.startsWith("editNewsButton")) {
-            // Find ID of news item to edit
-            let idToEdit = event.target.id.split("--")[1]
-            // Show edit form
-            let editFormHTML = editNewsForm(idToEdit)
-            printToDom(editFormHTML, "#newsFormSection")
-
-            // UPDATE button
-            document.querySelector("#updateNews").addEventListener("click", () => {
-                let news = document.querySelector("#editNewsTitle").value
-                let newsSynopsis = document.querySelector("#editNewsSynopsis").value
-                let newsURL = document.querySelector(`#newsLink--${idToEdit}`).value
-
-                let updatedNewsObject = {
-                    news: news,
-                    newsSynopsis: newsSynopsis,
-                    newsURL: newsURL,
-                    userId: parseInt(userId)
-                }
-
-                let apiString = `news/${idToEdit}`
-                // API call
-                API.EDIT(apiString, updatedNewsObject)
+            // Check that news belongs to active user
+            if (buttonParent.classList.contains(`user--${userId}`)) {
+                // Find ID of news item to remove
+                let idToDelete = event.target.id.split("--")[1]
+                let stringToDelete = `news/${idToDelete}`
+                // Delete from API
+                API.DELETE(stringToDelete)
+                    // Re-load news list
                     .then(response => {
                         createNews()
                     })
+            }
 
-                // Remove edit form
-                document.querySelector("#newsFormSection").innerHTML = ""
-            })
-    }
-})
+        // EDIT button
+        } else if (event.target.id.startsWith("editNewsButton")) {
+
+            // Check that news belongs to active user
+            if (buttonParent.classList.contains(`user--${userId}`)) {
+                // Find ID of news item to edit
+                let idToEdit = event.target.id.split("--")[1]
+                // Show edit form
+                let editFormHTML = editNewsForm(idToEdit)
+                printToDom(editFormHTML, "#newsFormSection")
+
+                // UPDATE button
+                document.querySelector("#updateNews").addEventListener("click", () => {
+                    let news = document.querySelector("#editNewsTitle").value
+                    let newsSynopsis = document.querySelector("#editNewsSynopsis").value
+                    let newsURL = document.querySelector(`#newsLink--${idToEdit}`).href
+
+                    let updatedNewsObject = {
+                        news: news,
+                        newsSynopsis: newsSynopsis,
+                        newsURL: newsURL,
+                        userId: parseInt(userId)
+                    }
+
+                    let apiString = `news/${idToEdit}`
+                    // API call
+                    API.EDIT(apiString, updatedNewsObject)
+                        .then(response => {
+                            createNews()
+                        })
+
+                    // Remove edit form
+                    document.querySelector("#newsFormSection").innerHTML = ""
+                })
+            }
+        }
+    })
 }
 
 export default newsEventListener
